\FanFerret\QuestionBundle\Entity\Question::$type
The Question entity now has an explicit $type member rather than \FanFerret\QuestionBundle\Controller\DefaultController::surveyAction depending on its JSON parameters having a "type" property.

This approach makes more sense since while we don't know exactly what form the parameters for a question may take on (as questions may be of heterogeneous types) we do know that a question will have a type.

Another advantage of this approach is that it will enable a much clearer and convenient SQL (DQL technically as we're using Doctrine) syntax in the future.  For example if you wanted to report on "open" type questions you wouldn't have to use JSON tricks or filter the results in PHP.