\FanFerret\QuestionBundle\Survey\Survey::render - Return \FanFerret\QuestionBundle\Utility\Renderable
Improves the consistency with which rendering is handled by the overall system: It's better to standardize on dealing with rendering using the \FanFerret\QuestionBundle\Utility\Renderable class everywhere rather than use it some places and use raw strings in others.

Also this is lazily evaluated, which is a plus if there's ever a chance the caller won't use the return value (for some reason).