Front-End - Text Box/Area Event Handling Fix
Events were being fired on the "change" event of text boxes/areas only.  This meant that the events were not fired until the user blurred the text box/area.  This didn't make for an experience as responsible, fluid, and intuitive as is ideal.  Accordingly bound to the "change" and "input" events of text boxes/areas.