Notification Change - Prevent Double Notification
When a SurveySession was created late (11 or more days after checkout) the next time the fanferretquestion:notification command was run both the first and second notification for that SurveySession would be sent.  This is clearly not what is intended by the notification system.

Reverted d177c1eddc018ca78e9ede782e90bfe1ad8b2cb6 and changed the fanferretquestion:notification command to send the second notification eight days after the first.  This is still theoretically/ideally 11 days after checkout but avoids spamming the guest with two notifications in the degenerate case.