Admin Front-End - Delivery - Bootstrap Markup
The delivery form is now rendered with Bootstrap markup including a column-based design (on "medium" screens and above).

Also now uses a date picker rather than allowing a date and time to be selected using drop downs et cetera (Shawn has stated that the ability to collect the time of checkout is unimportant).