FOSUserBundle Integration
The \FanFerret\QuestionBundle\Controller\AdminController class now expects the user to be logged in and represented by a \FanFerret\QuestionBundle\Entity\User instance.

Added the \FanFerret\QuestionBundle\Entity\Acl entity which links \FanFerret\QuestionBundle\Entity\User entities to \FanFerret\QuestionBundle\Entity\Survey entities in a given capacity (i.e. the manner in which the user is permitted to interact with the survey).  This allows for certain users to be allowed to both schedule deliveries and view all comment cards ("ROLE_ADMIN" on the survey) whereas others can only schedule deliveries (any other role).