\FanFerret\QuestionBundle\Entity\Group & \FanFerret\QuestionBundle\Entity\Property
This changes the way that admin routes/slugs work.  Previously slugs had one or two components: One for the slug, and an optional one for the group.  Now slugs have two or three components: One for the survey and one for the property, and an optional one for the group.

Admin ACLs can now delegate permissions on a per survey, per property, or per group basis.