localStorage - Token Acquisition Fix
Occasionally in FireFox survey.js was unable to acquire the survey token from the <script> tag which loaded it.  This meant that the token was undefined which caused it to be unable to load or store survey values.

Fixed this by increasing the specificity of the jQuery selector.  I have been unable to reproduce the issue with this change.