Custom Stylesheets
As per Shawn's request the "stylesheets" property of a Survey's entity parameters may now be an array of strings which give the URLs of custom CSS stylesheets which shall be included after all other default stylesheets.  If this property is missing no custom stylesheets are included and everything proceeds as before.