\FanFerret\QuestionBundle\Question\ConditionalNestedQuestion
Allows for nested questions to be answered conditionally based on a top level yes/no question.

For example: A user may be given a question which asks whether they ate at a particular restaurant.  If they answer yes they have to answer questions about that restaurant.  If they answer no they do not have to answer further questions (as it would not make sense to have them answer questions about a restaurant at which they did not eat).