\FanFerret\QuestionBundle\Entity\SurveySession - Single Button Screen
Marc wants the user interface for sending emails and creating SurveySession entities to be changed so that all the user must do is enter the guest's email address and submit.  This however leads to the inability to gather several pieces of information which are currently not nullable (and therefore assumed elsewhere not to be null) within the SurveySession entity:

-	Checkout time
-	First name
-	Last name
-	Room number

Accordingly these fields had to be made nullable and code elsewhere had to be changed to accomodate the fact that this information may now be missing.